refactor: undo move of helper to client
